requirements.txt
fastapi
uvicorn
pandas
openpyxl
pdfplumber
anthropic
python-dotenv
pydantic-settings
python-multipart
